Renames Interface to ServerInterface
